Fix looping over scenario keys

+++ scripts/create_empty_ts.py
@@ -32,7 +32,7 @@ from datetime import datetime
 from oemof_b3.config.config import load_yaml, settings
 from oemof_b3.model import model_structures
 from oemof_b3 import model
-from oemof_b3.tools.data_processing import HEADER_B3_TS, stack_timeseries
+from oemof_b3.tools.data_processing import HEADER_B3_TS, stack_timeseries, load_b3_timeseries
 
 
 def get_sub_dict(subsub_key, _dict):
@@ -182,6 +182,37 @@ def get_df_of_all_empty_ts(profile_names, _region):
     return ts_df
 
 
+def drop_duplicates(_df):
+    """
+        Remove duplicate rows from a pandas DataFrame based on specified columns.
+
+        Parameters
+        ----------
+        _df : pandas DataFrame
+            The DataFrame to remove duplicates from.
+
+        Returns
+        -------
+        _df : pandas DataFrame
+            The updated DataFrame with duplicate rows removed.
+
+        Notes
+        -----
+        Duplicate rows are determined based on the values in the specified columns. By default, all columns
+        except the "series" column are used to determine duplicates. If there are multiple rows with the same
+        values in the specified columns, only the first occurrence is kept and subsequent occurrences are
+        dropped.
+    """
+    columns = [col for col in _df.columns if col != "series"]
+
+    _df = _df.drop_duplicates(
+        subset=columns,
+        ignore_index=True,
+    )
+
+    return _df
+
+
 def save_ts(_df, path):
     """
     This function saves time series that contain values of datetime format
@@ -197,12 +228,16 @@ def save_ts(_df, path):
 
 
     """
-    _df.to_csv(
-        path,
-        index=True,
-        date_format="%Y-%m-%d %H:%M:%S",
-        sep=settings.general.separator,
-    )
+    if not os.path.exists(path):
+
+        _df.index.name = "id_ts"
+
+        _df.to_csv(
+            path,
+            index=True,
+            date_format="%Y-%m-%d %H:%M:%S",
+            sep=settings.general.separator,
+        )
 
 
 if __name__ == "__main__":
@@ -213,6 +248,10 @@ if __name__ == "__main__":
     path_empty_ts_efficiencies = sys.argv[4]
 
     scenarios = os.listdir(scenarios_dir)
+
+    all_load_ts = pd.DataFrame(columns=HEADER_B3_TS)
+    all_feedin_ts = pd.DataFrame(columns=HEADER_B3_TS)
+    all_efficiencies_ts = pd.DataFrame(columns=HEADER_B3_TS)
 
     for scenario_specs in scenarios:
         scenario_specs = load_yaml(os.path.join(scenarios_dir, scenario_specs))
@@ -243,15 +282,33 @@ if __name__ == "__main__":
             attr_subdict["efficiency"] for attr_subdict in foreign_keys_efficiency
         ]
 
+        ts_dict = {
+            "load": load_names,
+            "feedin": feedin_names,
+            "efficiency": efficiency_names,
+        }
+
         for region in model_structure["regions"]:
+
             if load_names:
                 df_loads = get_df_of_all_empty_ts(load_names, region)
-                save_ts(df_loads, path_empty_load_ts)
+                all_load_ts = pd.concat([all_load_ts, df_loads], ignore_index=True)
 
             if feedin_names:
                 df_feedin = get_df_of_all_empty_ts(feedin_names, region)
-                save_ts(df_feedin, path_empty_ts_feedin)
+                all_feedin_ts = pd.concat([all_feedin_ts, df_feedin], ignore_index=True)
 
             if efficiency_names:
                 df_efficiencies = get_df_of_all_empty_ts(efficiency_names, region)
-                save_ts(df_efficiencies, path_empty_ts_efficiencies)
+                all_efficiencies_ts = pd.concat([all_efficiencies_ts, df_efficiencies], ignore_index=True)
+
+    ts_dict = {
+        "load": (load_names, all_load_ts, path_empty_load_ts),
+        "feedin": (feedin_names, all_feedin_ts, path_empty_ts_feedin),
+        "efficiency": (efficiency_names, all_efficiencies_ts, path_empty_ts_efficiencies),
+    }
+
+    for ts_type, (ts_name, ts_df, ts_path) in ts_dict.items():
+        if ts_name:
+            drop_duplicates(ts_df)
+            save_ts(ts_df, ts_path)
